# backend/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CodeLand.io Backend API",
    "description": "API for user management, authentication, exercises, and scoring system",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@CodeLand.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Development server"
    },
    {
      "url": "https://api.codeland.io",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/exercises": {
      "get": {
        "tags": ["exercises"],
        "summary": "Get all exercises",
        "description": "Retrieve a list of all exercises",
        "responses": {
          "200": {
            "description": "Exercises retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "exercises": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Exercise"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["exercises"],
        "summary": "Add new exercise (Admin only)",
        "description": "Create a new exercise (requires admin privileges)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseCreate"
              },
              "example": {
                "title": "Sum Function",
                "body": "Define a function called sum that takes two integers as arguments and returns their sum.",
                "difficulty": 0,
                "test_cases": ["sum(2, 3)", "sum(5, 7)"],
                "solutions": ["5", "12"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exercise created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or insufficient privileges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/{exercise_id}": {
      "get": {
        "tags": ["exercises"],
        "summary": "Get exercise by ID",
        "description": "Retrieve a specific exercise by its ID",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Exercise ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Exercise"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Exercise not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["exercises"],
        "summary": "Update exercise (Admin only)",
        "description": "Update an existing exercise (requires admin privileges)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Exercise ID"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Updated Sum Function"
                  },
                  "body": {
                    "type": "string",
                    "example": "Updated description."
                  },
                  "difficulty": {
                    "type": "integer",
                    "example": 1
                  },
                  "test_cases": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["sum(2, 3)", "sum(5, 7)", "sum(1, 1)"]
                  },
                  "solutions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["5", "12", "2"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or insufficient privileges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Exercise not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/validate_code": {
      "post": {
        "tags": ["exercises"],
        "summary": "Validate user code",
        "description": "Validate user's code against an exercise's test cases",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "answer": {
                    "type": "string",
                    "example": "def sum(a, b):\n    return a + b"
                  },
                  "exercise_id": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": ["answer", "exercise_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code validation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "boolean"
                      },
                      "example": [true, true]
                    },
                    "user_results": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["5", "12"]
                    },
                    "all_correct": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data or compilation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Exercise not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Execution error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login or registration"
      }
    },
    "schemas": {
...
      "Exercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Sum Function"
          },
          "body": {
            "type": "string",
            "example": "Define a function called sum that takes two integers as arguments and returns their sum."
          },
          "difficulty": {
            "type": "integer",
            "example": 0,
            "description": "0: easy, 1: medium, 2: hard"
          },
          "test_cases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["sum(2, 3)", "sum(5, 7)"]
          },
          "solutions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["5", "12"]
          }
        },
        "required": ["id", "title", "body", "difficulty", "test_cases", "solutions"]
      },
      "ExerciseCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Sum Function"
          },
          "body": {
            "type": "string",
            "example": "Define a function called sum that takes two integers as arguments and returns their sum."
          },
          "difficulty": {
            "type": "integer",
            "example": 0,
            "description": "0: easy, 1: medium, 2: hard"
          },
          "test_cases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["sum(2, 3)", "sum(5, 7)"]
          },
          "solutions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["5", "12"]
          }
        },
        "required": ["title", "body", "difficulty", "test_cases", "solutions"]
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": ["status", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["fail", "error"],
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Invalid payload."
          }
        },
        "required": ["status", "message"]
      }
    }
  }
}